requirements.txt
Django>=4.2.2,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.26.3,<0.27
